Fixed casting 3

+++ src/modules/sensors/rc_update.cpp
@@ -526,15 +526,15 @@ RCUpdate::rc_poll(const ParameterHandles &parameter_handles)
 			/* Update parameters from RC Channels (tuning with RC) if activated */
 			if (hrt_elapsed_time(&_last_rc_to_param_map_time) > 1e6) {
 				// Edited by Alberto Ruiz Garcia: map to params, sinusoidal gain
-                double max_rate_gain = 0.0f;
-                double max_ff_gain = 0.0f;
+                float max_rate_gain = 0.0f;
+                float max_ff_gain = 0.0f;
                 param_get(param_find("FW_YR_PMAX"), &max_rate_gain);
                 param_get(param_find("FW_YR_FFMAX"), &max_ff_gain);
                 float param_1 = get_rc_value(rc_channels_s::RC_CHANNELS_FUNCTION_PARAM_1, 0.0, 1.0);
                 float param_2 = get_rc_value(rc_channels_s::RC_CHANNELS_FUNCTION_PARAM_2, 0.0, 1.0);
                 // Compute gains
-                float yaw_rate_gain_new = max_rate_gain * sin(M_PI_2 * (double)param_1);
-                float yaw_feedforward_gain_new = max_ff_gain * sin(M_PI_2 * (double)param_2);
+                float yaw_rate_gain_new = (double)max_rate_gain * sin(M_PI_2 * (double)param_1);
+                float yaw_feedforward_gain_new =(double)max_ff_gain * sin(M_PI_2 * (double)param_2);
 
                 // Set parameters
                 if (yaw_rate_gain_new > _yaw_rate_gain + _yaw_rate_gain_sensitivity ||
